divide to pos interface components
